+++ relevanceai/dataset_api/dataset.py
@@ -3,41 +3,74 @@ Pandas like dataset API
 """
 import warnings
 import pandas as pd
+from relevanceai.dataset_api.groupby import Groupby, Agg
+from relevanceai.dataset_api.centroids import Centroids
+from typing import List, Union, Callable
+import math
 
 from relevanceai.vector_tools.client import VectorTools
 from relevanceai.api.client import BatchAPIClient
-from typing import List, Union
 
 
-class Series:
+class Series(BatchAPIClient):
     """
     A wrapper class for being able to vectorize documents over field
     """
 
-    def __init__(self, client: BatchAPIClient) -> None:
+    def __init__(self, project: str, api_key: str, dataset_id: str, field) -> None:
         """
         Initialise the class
         """
-        self.client = client
+        self.project = project
+        self.api_key = api_key
+        self.dataset_id = dataset_id
+        self.field = field
+        super().__init__(project=project, api_key=api_key)
 
-    def __call__(self, dataset_id: str, field: str):
+    def sample(
+        self, n: int = 0, frac: float = None, filters: list = [], random_state: int = 0
+    ):
         """
-        Instaniates a Series
+        Return a random sample of items from a dataset.
 
         Parameters
         ----------
-        dataset_id : str
-            The dataset_id of concern
-        field : str
-            The field within the dataset that you would like to select
+        n : int
+            Number of items to return. Cannot be used with frac.
+        frac: float
+            Fraction of items to return. Cannot be used with n.
+        filters: list
+            Query for filtering the search results
+        random_state: int
+            Random Seed for retrieving random documents.
 
-        Returns
-        -------
-        Self
         """
-        self.dataset_id = dataset_id
-        self.field = field
-        return self
+        select_fields = [self.field] if isinstance(self.field, str) else self.field
+        return Dataset(self.project, self.api_key)(self.dataset_id).sample(
+            n=n,
+            frac=frac,
+            filters=filters,
+            random_state=random_state,
+            select_fields=select_fields,
+        )
+
+    def all(
+        self,
+        chunk_size: int = 1000,
+        filters: List = [],
+        sort: List = [],
+        include_vector: bool = True,
+        show_progress_bar: bool = True,
+    ):
+        select_fields = [self.field] if isinstance(self.field, str) else self.field
+        return Dataset(self.project, self.api_key)(self.dataset_id).all(
+            chunk_size=chunk_size,
+            filters=filters,
+            sort=sort,
+            select_fields=select_fields,
+            include_vector=include_vector,
+            show_progress_bar=show_progress_bar,
+        )
 
     def vectorize(self, model) -> None:
         """
@@ -59,6 +92,62 @@ class Series:
                 return model([self.field], documents)
 
         self.client.pull_update_push(self.dataset_id, encode_documents)
+
+    def apply(
+        self,
+        func: Callable,
+        axis: int = 0,
+    ):
+        """
+        Apply a function along an axis of the DataFrame.
+
+        Objects passed to the function are Series objects whose index is either the DataFrame’s index (axis=0) or the DataFrame’s columns (axis=1). By default (result_type=None), the final return type is inferred from the return type of the applied function. Otherwise, it depends on the result_type argument.
+
+        Parameters
+        --------------
+        func: function
+            Function to apply to each document
+
+        axis: int
+            Axis along which the function is applied.
+            - 9 or 'index': apply function to each column
+            - 1 or 'columns': apply function to each row
+
+        Example
+        ---------------
+
+        >>> df["sample_1_label"].apply(lambda x: x + 3)
+
+        """
+        if axis == 1:
+            raise ValueError("We do not support column-wise operations!")
+
+        def bulk_fn(documents):
+            for d in documents:
+                try:
+                    if self.is_field(self.field, d):
+                        self.set_field(
+                            self.field, d, func(self.get_field(self.field, d))
+                        )
+                except Exception as e:
+                    continue
+            return documents
+
+        return self.pull_update_push(
+            self.dataset_id, bulk_fn, select_fields=[self.field]
+        )
+
+    def __getitem__(self, loc: Union[int, str]):
+        if isinstance(loc, int):
+            warnings.warn(
+                "Integer selection of dataframe is not stable at the moment. Please use a string ID if possible to ensure exact selection."
+            )
+            return self.get_documents(
+                self.dataset_id, loc + 1, select_fields=[self.field]
+            )[loc][self.field]
+        elif isinstance(loc, str):
+            return self.datasets.documents.get(self.dataset_id, loc)[self.field]
+        raise TypeError("Incorrect data type! Must be a string or an integer")
 
 
 class Dataset(BatchAPIClient):
@@ -110,6 +199,9 @@ class Dataset(BatchAPIClient):
         self.audio_fields = audio_fields
         self.highlight_fields = highlight_fields
         self.output_format = output_format
+        self.groupby = Groupby(self.project, self.api_key, self.dataset_id)
+        self.agg = Agg(self.project, self.api_key, self.dataset_id)
+        self.centroids = Centroids(self.project, self.api_key, self.dataset_id)
 
         return self
 
@@ -129,13 +221,13 @@ class Dataset(BatchAPIClient):
         n_documents = self.get_number_of_documents(dataset_id=self.dataset_id)
         return (n_documents, len(schema))
 
-    def __getitem__(self, field: str):
+    def __getitem__(self, field):
         """
         Returns a Series Object that selects a particular field within a dataset
 
         Parameters
         ----------
-        field : str
+        field
             the particular field within the dataset
 
         Returns
@@ -143,36 +235,57 @@ class Dataset(BatchAPIClient):
         Tuple
             (N, C)
         """
-        series = Series(self)
-        series(dataset_id=self.dataset_id, field=field)
-        return series
+        return Series(self.project, self.api_key, self.dataset_id, field)
 
-    def info(self) -> dict:
+    def _get_possible_dtypes(self, schema):
+        possible_dtypes = []
+        for v in schema.values():
+            if isinstance(v, str):
+                possible_dtypes.append(v)
+            elif isinstance(v, dict):
+                if list(v)[0] == "vector":
+                    possible_dtypes.append("vector_")
+        return possible_dtypes
+
+    def _get_dtype_count(self, schema: dict):
+        possible_dtypes = self._get_possible_dtypes(schema)
+        dtypes = {
+            dtype: list(schema.values()).count(dtype) for dtype in possible_dtypes
+        }
+        return dtypes
+
+    def _get_schema(self):
+        # stores schema in memory to save users API usage/reloading
+        if hasattr(self, "_schema"):
+            return self._schema
+        self._schema = self.datasets.schema(self.dataset_id)
+        return self._schema
+
+    def info(self, dtype_count: bool = False) -> pd.DataFrame:
         """
         Return a dictionary that contains information about the Dataset
         including the index dtype and columns and non-null values.
 
+        Parameters
+        -----------
+        dtype_count: bool
+            If dtype_count is True, prints a value_counts of the data type
+
+
         Returns
-        -------
+        ---------
         Dict
             Dictionary of information
         """
-        health = self.datasets.monitor.health(self.dataset_id)
-        schema = self.datasets.schema(self.dataset_id)
-        schema = {key: str(value) for key, value in schema.items()}
-        info = {
-            key: {
-                "Non-Null Count": health[key]["missing"],
-                "Dtype": schema[key],
-            }
-            for key in schema.keys()
-        }
-        dtypes = {
-            dtype: list(schema.values()).count(dtype)
-            for dtype in set(list(schema.values()))
-        }
-        info = {"info": info, "dtypes": dtypes}
-        return info
+        health: dict = self.datasets.monitor.health(self.dataset_id)
+        schema: dict = self._get_schema()
+        info_df = pd.DataFrame()
+        info_df["Non-Null Count"] = [health[key]["missing"] for key in schema]
+        info_df["Dtype"] = [schema[key] for key in schema]
+        if dtype_count:
+            dtypes_info = self._get_dtype_count(schema)
+            print(dtypes_info)
+        return info_df
 
     def head(
         self, n: int = 5, raw_json: bool = False, **kw
@@ -265,6 +378,122 @@ class Dataset(BatchAPIClient):
             overwrite=overwrite,
         )
         return centroids
+
+    def sample(
+        self,
+        n: int = 0,
+        frac: float = None,
+        filters: list = [],
+        random_state: int = 0,
+        select_fields: list = [],
+    ):
+
+        """
+        Return a random sample of items from a dataset.
+
+        Parameters
+        ----------
+        n : int
+            Number of items to return. Cannot be used with frac.
+        frac: float
+            Fraction of items to return. Cannot be used with n.
+        filters: list
+            Query for filtering the search results
+        random_state: int
+            Random Seed for retrieving random documents.
+        select_fields: list
+            Fields to include in the search results, empty array/list means all fields.
+
+        """
+        if n == 0 and frac is None:
+            raise ValueError("Must provide one of n or frac")
+
+        if frac and n:
+            raise ValueError("Only one of n or frac can be provided")
+
+        if frac:
+            if frac > 1 or frac < 0:
+                raise ValueError("Fraction must be between 0 and 1")
+            n = math.ceil(
+                self.get_number_of_documents(self.dataset_id, filters=filters) * frac
+            )
+
+        return self.datasets.documents.get_where(
+            dataset_id=self.dataset_id,
+            filters=filters,
+            page_size=n,
+            random_state=random_state,
+            is_random=True,
+            select_fields=select_fields,
+        )["documents"]
+
+    def apply(
+        self,
+        func: Callable,
+        axis: int = 0,
+    ):
+        """
+        Apply a function along an axis of the DataFrame.
+
+        Objects passed to the function are Series objects whose index is either the DataFrame’s index (axis=0) or the DataFrame’s columns (axis=1). By default (result_type=None), the final return type is inferred from the return type of the applied function. Otherwise, it depends on the result_type argument.
+
+        Parameters
+        --------------
+        func: function
+            Function to apply to each document
+
+        axis: int
+            Axis along which the function is applied.
+            - 9 or 'index': apply function to each column
+            - 1 or 'columns': apply function to each row
+
+        """
+        if axis == 1:
+            raise ValueError("We do not support column-wise operations!")
+
+        def bulk_fn(docs):
+            for d in docs:
+                func(d)
+            return docs
+
+        return self.pull_update_push(self.dataset_id, bulk_fn)
+
+    def all(
+        self,
+        chunk_size: int = 1000,
+        filters: List = [],
+        sort: List = [],
+        select_fields: List = [],
+        include_vector: bool = True,
+        show_progress_bar: bool = True,
+    ):
+
+        """
+        Retrieve all documents with filters. Filter is used to retrieve documents that match the conditions set in a filter query. This is used in advance search to filter the documents that are searched. For more details see documents.get_where.
+
+        Parameters
+        ----------
+        chunk_size : list
+            Number of documents to retrieve per retrieval
+        include_vector: bool
+            Include vectors in the search results
+        sort: list
+            Fields to sort by. For each field, sort by descending or ascending. If you are using descending by datetime, it will get the most recent ones.
+        filters: list
+            Query for filtering the search results
+        select_fields : list
+            Fields to include in the search results, empty array/list means all fields.
+        """
+
+        return self.get_all_documents(
+            dataset_id=self.dataset_id,
+            chunk_size=chunk_size,
+            filters=filters,
+            sort=sort,
+            select_fields=select_fields,
+            include_vector=include_vector,
+            show_progress_bar=show_progress_bar,
+        )
 
 
 class Datasets(BatchAPIClient):